remove the comments and code for dynamically loading ioredis
